requirements.txt
python-telegram-bot==20.8
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
urllib3==2.2.1 